refactor(pipeline): leaf the retrieval key and the two entry reads
generation-context imported the whole retrieval phase — runRetrieval, the embedder, the swap manager, the probe — to read one string literal off it, which is module-load cost every context builder and its tests pay, and a cycle the moment retrieval wants a read the builder owns; that moment is the next commit but one, so both the key and the scene/lastTurns queries move to leaves nothing points back at.
